+++ labor-spring-core/src/main/java/com/labor/spring/feign/auth/AuthLoginService.java
@@ -26,7 +26,7 @@ public interface AuthLoginService {
 	public User fetchUser(String accessToken);
 	public String fetchUserToken(String type,String code);
 	
-	public boolean isCurrentUserOrSuperUser(Integer userid, String useruuid);
+	public boolean isCurrentUserOrSuperUser(Long userid, String useruuid);
 	
 	public String create(
 					String clientKey,
