+++ labor-spring-system-auth/src/main/java/com/labor/spring/core/service/PermissionServiceIntf.java
@@ -10,12 +10,12 @@ public interface PermissionServiceIntf {
 
 	public Integer initialization();
 	public List<Permission> findListByStatus(String status);
-	public List<Permission> findListByRoleid(Integer roleid);
-	public List<Permission> findListByUserid(Integer userid);
+	public List<Permission> findListByRoleid(Long roleid);
+	public List<Permission> findListByUserid(Long userid);
 	
 	public List<Permission> findListByStatus(String status, String type);
-	public List<Permission> findListByRoleid(Integer roleid, String type);
-	public List<Permission> findListByUserid(Integer userid, String type);
+	public List<Permission> findListByRoleid(Long roleid, String type);
+	public List<Permission> findListByUserid(Long userid, String type);
 	
 	public Permission save(Permission permission);
 	public void refreshByType(Set<String> permissions, String type);
