+++ labor-spring-system-ppp/src/main/java/com/labor/spring/system/ppp/api/document/DocumentRestController.java
@@ -290,7 +290,7 @@ public class DocumentRestController {
 			ret.getDocument().setFilePath(hidepath+"hasdownload|");
 		}
 		User creator = ret.getCreator();
-		Integer creatorid = 0;
+		Long creatorid = 0L;
 		if (creator!=null) {
 			creatorid = creator.getId();
 		}
